requirements.txt
python-telegram-bot==20.6
aiosqlite==0.19.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.0
asyncio==3.4.3
tabulate==0.9.0
aiohttp==3.9.1
typing-extensions>=4.0.0
wheel>=0.40.0
setuptools>=65.5.1
radix-engine-toolkit==2.0.0 